Added basic dishes list page, improved filtering on back-end
Dishes list pagination and filtering need to be implemented. Also I removed some dishes cards elements which are not yet implemented. Filtering on backend was modified to display the total number of elements in a collection, the number of elements matching filters and the number of resulting pages when pagination is used. Code needs to be improved - I will work on it later.